review: split pin detector into strategy partials, guard port-shape chip, remove agent tooling files
- GdsPinDetector.cs is now a thin orchestrator (133 lines); label/marker
  detection moved to GdsPinDetector.LabelPins.cs and the edge heuristic
  (incl. fallback) to GdsPinDetector.EdgeScan.cs — all partials <=250 lines
- AddPortShapeLayers skips pairs already claimed waveguide/metal so a
  through-going core strip no longer gets a "may mark ports" chip (+ test)
- off-center port-shape fixture pins the GDS->app Y-flip conversion
- untrack .codegraph/ agent tooling files (unrelated to this PR)
