--- .codegraph/config.json
@@ -1,143 +1,0 @@
-{
-  "version": 1,
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    "**/*.js",
-    "**/*.jsx",
-    "**/*.py",
-    "**/*.go",
-    "**/*.rs",
-    "**/*.java",
-    "**/*.c",
-    "**/*.h",
-    "**/*.cpp",
-    "**/*.hpp",
-    "**/*.cc",
-    "**/*.cxx",
-    "**/*.cs",
-    "**/*.php",
-    "**/*.rb",
-    "**/*.swift",
-    "**/*.kt",
-    "**/*.kts",
-    "**/*.dart",
-    "**/*.svelte",
-    "**/*.vue",
-    "**/*.liquid",
-    "**/*.pas",
-    "**/*.dpr",
-    "**/*.dpk",
-    "**/*.lpr",
-    "**/*.dfm",
-    "**/*.fmx",
-    "**/*.scala",
-    "**/*.sc"
-  ],
-  "exclude": [
-    "**/.git/**",
-    "**/node_modules/**",
-    "**/vendor/**",
-    "**/Pods/**",
-    "**/dist/**",
-    "**/build/**",
-    "**/out/**",
-    "**/bin/**",
-    "**/obj/**",
-    "**/target/**",
-    "**/*.min.js",
-    "**/*.bundle.js",
-    "**/.next/**",
-    "**/.nuxt/**",
-    "**/.svelte-kit/**",
-    "**/.output/**",
-    "**/.turbo/**",
-    "**/.cache/**",
-    "**/.parcel-cache/**",
-    "**/.vite/**",
-    "**/.astro/**",
-    "**/.docusaurus/**",
-    "**/.gatsby/**",
-    "**/.webpack/**",
-    "**/.nx/**",
-    "**/.yarn/cache/**",
-    "**/.pnpm-store/**",
-    "**/storybook-static/**",
-    "**/.expo/**",
-    "**/web-build/**",
-    "**/ios/Pods/**",
-    "**/ios/build/**",
-    "**/android/build/**",
-    "**/android/.gradle/**",
-    "**/__pycache__/**",
-    "**/.venv/**",
-    "**/venv/**",
-    "**/site-packages/**",
-    "**/dist-packages/**",
-    "**/.pytest_cache/**",
-    "**/.mypy_cache/**",
-    "**/.ruff_cache/**",
-    "**/.tox/**",
-    "**/.nox/**",
-    "**/*.egg-info/**",
-    "**/.eggs/**",
-    "**/go/pkg/mod/**",
-    "**/target/debug/**",
-    "**/target/release/**",
-    "**/.gradle/**",
-    "**/.m2/**",
-    "**/generated-sources/**",
-    "**/.kotlin/**",
-    "**/.dart_tool/**",
-    "**/.vs/**",
-    "**/.nuget/**",
-    "**/artifacts/**",
-    "**/publish/**",
-    "**/cmake-build-*/**",
-    "**/CMakeFiles/**",
-    "**/bazel-*/**",
-    "**/vcpkg_installed/**",
-    "**/.conan/**",
-    "**/Debug/**",
-    "**/Release/**",
-    "**/x64/**",
-    "**/.pio/**",
-    "**/release/**",
-    "**/*.app/**",
-    "**/*.asar",
-    "**/DerivedData/**",
-    "**/.build/**",
-    "**/.swiftpm/**",
-    "**/xcuserdata/**",
-    "**/Carthage/Build/**",
-    "**/SourcePackages/**",
-    "**/__history/**",
-    "**/__recovery/**",
-    "**/*.dcu",
-    "**/.composer/**",
-    "**/storage/framework/**",
-    "**/bootstrap/cache/**",
-    "**/.bundle/**",
-    "**/tmp/cache/**",
-    "**/public/assets/**",
-    "**/public/packs/**",
-    "**/.yardoc/**",
-    "**/coverage/**",
-    "**/htmlcov/**",
-    "**/.nyc_output/**",
-    "**/test-results/**",
-    "**/.coverage/**",
-    "**/.idea/**",
-    "**/logs/**",
-    "**/tmp/**",
-    "**/temp/**",
-    "**/_build/**",
-    "**/docs/_build/**",
-    "**/site/**"
-  ],
-  "languages": [],
-  "frameworks": [],
-  "maxFileSize": 1048576,
-  "extractDocstrings": true,
-  "trackCallSites": true
-}
